requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
torchvision